Fix bug#39: AFS-current order doesn't display correctly
Changes: view current order will be queried base on the local of the
default address of the signed in consumer. If the consumer does not have
a default address, it will use the  local time of the first address. If
consumer has no addresses, it will query base on the machine time